organise tests of various closures
