+++ test/runtests.jl
@@ -22,8 +22,8 @@ end
   #  include("patch.jl")
 
     include("concept.jl")
-    include("scope.jl")
-    #include("closure.jl")
+    include("closure.jl")
+    #include("scope.jl")
     #include("import.jl")
     include("real-open.jl")
     include("real-isfile.jl")
